+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Added
 - SC2259/SC2260: Warn when redirections override pipes
 - SC2261: Warn about multiple competing redirections
+- SC2262/SC2263: Warn about aliases declared and used in the same parsing unit
 
 ### Fixed
 - SC1072/SC1073 now respond to disable annotations, though ignoring parse errors